requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
psycopg2-binary==2.9.9
Authlib==1.3.0
python-dotenv==1.0.1
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
weasyprint==60.2
reportlab==4.2.0